refactor: 검증 로직 utils 폴더로 이동
